Move extensions and fix issues
