+++ GDEdit/GDEdit/Utilities/Functions/General/UsefulFunctions.cs
@@ -5,7 +5,7 @@ using System.Text;
 using System.Threading.Tasks;
 using static System.IO.File;
 
-namespace EffectSome.Utilities.Functions.General
+namespace GDEdit.Utilities.Functions.General
 {
     public static class UsefulFunctions
     {
